feat: added active devices metric api
